* add: logout system and register

+++ client/src/pages/Register.tsx
@@ -7,9 +7,10 @@ import { useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
 
 interface LoginForm {
+    name: string;
     email: string;
     password: string;
-    remember: boolean;
+    confirm_password: string;
 }
 
 interface LoginData {
@@ -23,17 +24,16 @@ interface LoginData {
 axios.defaults.withCredentials = true;
 axios.defaults.baseURL = 'http://localhost:8000/api/';
 
-export default function Login() {
-    const [loginForm, setLoginForm] = useState<LoginForm>({ email: '', password: '' , remember: false});
+export default function Register() {
+    const [loginForm, setLoginForm] = useState<LoginForm>({ email: '', name: '', password: '' , confirm_password: ''});
     const [error, setError] = useState<string | null>(null);
     const navigate = useNavigate();
 
     const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        const { name, type, value, checked } = event.target;
+        const { name, value } = event.target;
     
         setLoginForm(prev => ({
-            ...prev,
-            [name]: type === 'checkbox' ? checked : value  
+            ...prev, [name]: value  
         }));
     };
     
@@ -41,13 +41,15 @@ export default function Login() {
     const handleSubmit = async (event: React.FormEvent) => {
         event.preventDefault();
         try {
-            const response = await axios.post<LoginData>('login', loginForm);
-            const token = response.data.token;
-            const user = response.data.user;
-            console.log(token)
-            localStorage.setItem('authToken', token);
-            localStorage.setItem('user', JSON.stringify(user));
-            navigate('/');
+            if (loginForm.password == loginForm.confirm_password) {
+                const response = await axios.post<LoginData>('register', loginForm);
+                const token = response.data.token;
+                const user = response.data.user;
+                console.log(token)
+                localStorage.setItem('authToken', token);
+                localStorage.setItem('user', JSON.stringify(user));
+                navigate('/');
+            }
         } catch {
             setError('Dados incorretos')
         }
@@ -57,6 +59,16 @@ export default function Login() {
         <AuthLayout>
             <form onSubmit={handleSubmit} className="flex flex-col gap-4 p-10 rounded-md shadow-md bg-zinc-100">
                 {error ? <p className="text-red-500">{error}</p> : ''}
+                <div className="flex flex-col">
+                    <Label>Name:</Label>
+                    <input
+                        type="string"
+                        className="px-2 py-1 text-blue-600 bg-transparent border rounded-lg border-sm border-sky-700"
+                        name="name"
+                        value={loginForm.name}
+                        onChange={handleInputChange}
+                    />
+                </div>
                 <div className="flex flex-col">
                     <Label>Email:</Label>
                     <input
@@ -77,18 +89,18 @@ export default function Login() {
                         onChange={handleInputChange}
                     />
                 </div>
-                <div className="flex items-center gap-4">
+                <div className="flex flex-col">
+                    <Label>Confirmar Senha:</Label>
                     <input
-                        type="checkbox"
-                        className="w-4 h-4 border-2 border-blue-500 rounded-full bg-blue checked:bg-blue-500 checked:border-blue-500 focus:ring-2 focus:ring-blue-500 focus:outline-none"
-                        name="remember"
-                        checked={loginForm.remember}
+                        type="password"
+                        className="px-2 py-1 text-blue-600 bg-transparent border rounded-lg border-sm border-sky-700"
+                        name="confirm_password"
+                        value={loginForm.confirm_password}
                         onChange={handleInputChange}
-                        />
-                    <Label>Lembrar Sessão?</Label>
+                    />
                 </div>
-                <ButtonPrimary>Logar</ButtonPrimary>
-                <Link className="underline text-sky-500 hover:text-sky-600" to={'/register'}>Não tem uma conta? Cadastrar-se!</Link>
+                <ButtonPrimary>Registrar</ButtonPrimary>
+                <Link className="underline text-sky-500 hover:text-sky-600" to={'/login'}>Já tem uma conta? Logar-se!</Link>
             </form>
         </AuthLayout>
     );
